Switch vanilla example to instruction-tuned Gemma-4; sweeps/trainer tidy-ups
- examples/train_vanilla_sae.py: train on layer-14 input of
  instruction-tuned Gemma-4 E2B over lmsys-chat-1m conversations
  (was GPT-2 block 6 / openwebtext); d_data -> GEMMA_4_DEFAULT_D_DATA;
  betas/use_lars tuned; docstring updated incl. HF-gating + the
  d_data-placeholder caveat.
- sweeps/{do_sweep,newsweeper,sweeper}.py: drop a dead comment;
  remove unused TypeVar; narrow None via asserts + cast(SweepFile, …)
  so the spec/sweep_data accesses are well-typed.
- trainer.py: drop two stale "maybe cast other direction" TODOs.
- sae_architecture.py: collapse a one-arg signature; keep the
  dataloader-naming TODO as tracked backlog.

ruff check (full CI scope): All checks passed. 204 tests pass.

+++ src/saeco/trainer/trainer.py
@@ -251,8 +251,8 @@ class Trainer:
             x: SAETrainBatch
             inp, target = x.input, x.target
             if not self.cfg.use_autocast:
-                inp = inp.float()  # TODO maybe cast other direction instead
-                target = target.float()  # TODO maybe cast other direction instead
+                inp = inp.float()
+                target = target.float()
             self.optim.zero_grad()
             if self.t % self.eval_step_freq == 0 or (
                 self.cfg.schedule.run_length
